requirements.txt
absl-py==0.4.1
astor==0.7.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2018.8.24
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==1.2.0
Markdown==2.6.11
MarkupSafe==1.0
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.11.3
olefile==0.46
opencv-python==4.1.2.30
Pillow==5.2.0
protobuf==3.6.0
PyYAML==3.13
scikit-learn==0.18
scipy==0.18.1
six==1.11.0
TBB==0.1
tensorflow==1.2.1
termcolor==1.1.0
Theano==0.8.2
Werkzeug==0.16.0
wincertstore==0.2
